requirements.txt
Flask==2.3.2
urllib3==1.26.18
Pillow==9.4.0